+++ CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- ChkTeX is now opt-in instead of running by default.
+
 ### Removed
 
 ## [0.2.0] - 2026-09-14